# backend/openapi.yaml
openapi: 3.1.0
info:
  title: Budget App API
  version: 1.0.0
  description: REST API for the budget tracking application
servers:
  - url: http://localhost:5000
    description: Local development
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Category:
      type: object
      required: [id, name, color, budgetAmount, spendLimit, sortOrder]
      properties:
        id: { type: integer }
        name: { type: string }
        color: { type: string }
        budgetAmount: { type: number }
        spendLimit: { type: number }
        sortOrder: { type: integer }
...
    RecurringTransaction:
      type: object
      required: [id, name, amount, type, categoryId, dayOfMonth, isActive, createdAt, updatedAt]
      properties:
        id: { type: integer }
        name: { type: string }
        amount: { type: number }
        type:
          type: string
          enum: [expense, income]
        categoryId: { type: integer }
        dayOfMonth: { type: integer, minimum: 1, maximum: 31 }
        isActive: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    Account:
      type: object
      required: [id, name, color, accountType, initialBalance, createdAt, updatedAt]
      properties:
        id: { type: integer }
        name: { type: string }
        color: { type: string }
        accountType:
          type: string
          enum: [checking, savings, brokerage, '401k', ira, roth_ira, hsa, other]
        initialBalance: { type: number }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    MonthBudget:
      type: object
      required: [monthKey, takeHomePay, categories, isLocked, createdAt, updatedAt]
      properties:
        monthKey: { type: string, example: '2026-03' }
        takeHomePay: { type: number }
        categories:
          type: array
          items: { $ref: '#/components/schemas/Category' }
        isLocked: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    AuthTokens:
      type: object
      required: [accessToken, refreshToken]
      properties:
        accessToken: { type: string }
        refreshToken: { type: string }
    Error:
      type: object
      required: [message]
      properties:
        message: { type: string }
        code: { type: string }
security:
  - bearerAuth: []
paths:
...
  /api/auth/register:
    post:
      tags: [Auth]
      summary: Register a new user account
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthTokens' }
        '409':
          description: Email already in use
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Log in and receive tokens
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthTokens' }
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /api/auth/logout:
    post:
      tags: [Auth]
      summary: Invalidate refresh token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        '204':
          description: Logged out
  /api/auth/refresh:
    post:
      tags: [Auth]
      summary: Exchange refresh token for new access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        '200':
          description: New access token issued
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthTokens' }
        '401':
          description: Refresh token invalid or expired
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  # ─── Settings ──────────────────────────────────────────────────────────────
...
  /api/months/{monthKey}:
    parameters:
      - name: monthKey
        in: path
        required: true
        schema: { type: string, example: '2026-03' }
    get:
      tags: [Months]
      summary: Get a single month budget
      responses:
        '200':
          description: Month budget
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MonthBudget' }
        '404':
          description: Month not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /api/months/{monthKey}/init:
    parameters:
      - name: monthKey
        in: path
        required: true
        schema: { type: string, example: '2026-03' }
    post:
      tags: [Months]
      summary: Initialize a month from templates and recurring transactions
      responses:
        '201':
          description: Initialized month budget
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MonthBudget' }
        '409':
          description: Month already initialized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /api/months/{monthKey}/income:
    parameters:
      - name: monthKey
        in: path
        required: true
        schema: { type: string, example: '2026-03' }
    put:
      tags: [Months]
      summary: Update take-home pay for a month
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [takeHomePay]
              properties:
                takeHomePay: { type: number }
      responses:
        '200':
          description: Updated month budget
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MonthBudget' }
        '404':
          description: Month not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /api/months/{monthKey}/lock:
    parameters:
      - name: monthKey
        in: path
        required: true
        schema: { type: string, example: '2026-03' }
    post:
      tags: [Months]
      summary: Lock a month (prevents further edits)
      responses:
        '200':
          description: Locked month budget
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MonthBudget' }
        '404':
          description: Month not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  # ─── Categories ────────────────────────────────────────────────────────────
  /api/months/{monthKey}/categories:
    parameters:
      - name: monthKey
        in: path
        required: true
        schema: { type: string, example: '2026-03' }
    get:
      tags: [Categories]
      summary: List categories for a month
      responses:
        '200':
          description: Categories
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Category' }
    post:
      tags: [Categories]
      summary: Add a category to a month
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, color, budgetAmount, spendLimit, sortOrder]
              properties:
                name: { type: string }
                color: { type: string }
                budgetAmount: { type: number }
                spendLimit: { type: number }
                sortOrder: { type: integer }
      responses:
        '201':
          description: Created category
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Category' }
  /api/months/{monthKey}/categories/{id}:
    parameters:
      - name: monthKey
        in: path
        required: true
        schema: { type: string }
      - name: id
        in: path
        required: true
        schema: { type: integer }
    put:
      tags: [Categories]
      summary: Update a category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                color: { type: string }
                budgetAmount: { type: number }
                spendLimit: { type: number }
                sortOrder: { type: integer }
      responses:
        '200':
          description: Updated category
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Category' }
        '404':
          description: Category not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
    delete:
      tags: [Categories]
      summary: Delete a category
      responses:
        '204':
          description: Deleted
        '404':
          description: Category not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  # ─── Transactions ──────────────────────────────────────────────────────────
...
  /api/recurring:
    get:
      tags: [Recurring]
      summary: List all recurring transactions
      responses:
        '200':
          description: List of recurring transactions
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/RecurringTransaction' }
    post:
      tags: [Recurring]
      summary: Create a recurring transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, amount, type, categoryId, dayOfMonth, isActive]
              properties:
                name: { type: string }
                amount: { type: number }
                type:
                  type: string
                  enum: [expense, income]
                categoryId: { type: integer }
                dayOfMonth: { type: integer, minimum: 1, maximum: 31 }
                isActive: { type: boolean }
      responses:
        '201':
          description: Created recurring transaction
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RecurringTransaction' }
  /api/recurring/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: integer }
    put:
      tags: [Recurring]
      summary: Update a recurring transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                amount: { type: number }
                type:
                  type: string
                  enum: [expense, income]
                categoryId: { type: integer }
                dayOfMonth: { type: integer, minimum: 1, maximum: 31 }
                isActive: { type: boolean }
      responses:
        '200':
          description: Updated recurring transaction
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RecurringTransaction' }
        '404':
          description: Not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
    delete:
      tags: [Recurring]
      summary: Delete a recurring transaction
      responses:
        '204':
          description: Deleted
        '404':
          description: Not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  # ─── Accounts ──────────────────────────────────────────────────────────────
  /api/accounts:
    get:
      tags: [Accounts]
      summary: List all accounts
      responses:
        '200':
          description: List of accounts
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Account' }
    post:
      tags: [Accounts]
      summary: Create an account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, color, accountType, initialBalance]
              properties:
                name: { type: string }
                color: { type: string }
                accountType:
                  type: string
                  enum: [checking, savings, brokerage, '401k', ira, roth_ira, hsa, other]
                initialBalance: { type: number }
      responses:
        '201':
          description: Created account
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Account' }
  /api/accounts/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: integer }
    put:
      tags: [Accounts]
      summary: Update an account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                color: { type: string }
                accountType:
                  type: string
                  enum: [checking, savings, brokerage, '401k', ira, roth_ira, hsa, other]
                initialBalance: { type: number }
      responses:
        '200':
          description: Updated account
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Account' }
        '404':
          description: Account not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
    delete:
      tags: [Accounts]
      summary: Delete an account
      responses:
        '204':
          description: Deleted
        '404':
          description: Account not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }